+++ src/components/layouts/index.ts
@@ -1,0 +1,5 @@
+export { default as ArticleLayout } from "./ArticleLayout";
+export { default as DefaultLayout } from "./DefaultLayout";
+export { default as HomeLayout } from "./HomeLayout";
+export { RootLayout } from "./RootLayout";
+export { TwoColumnLayout } from "./TwoColumnLayout";
