+++ src/components/pages/articles/ArticleDetail/index.tsx
@@ -7,7 +7,7 @@ import MDXCustomComponents from "@/components/atoms/mdx";
 import TextDate from "@/components/atoms/texts/TextDate";
 import Thumbnail from "@/components/atoms/Thumbnail";
 import { HtmlHeadBase, HtmlHeadJsonLd } from "@/components/functions/meta";
-import ArticleLayout from "@/components/layouts/ArticleLayout";
+import { ArticleLayout } from "@/components/layouts";
 import { TagListColored } from "@/components/molecules/TagList";
 import { getSafeDate } from "@/lib/date";
 import type { TArticle, TCategory, TConfig, TPickup, TRankedArticle } from "@/types";
